requirements.txt
pytesseract==0.4.0
Pillow==8.3.2
scikit-learn==1.2.2
numpy==1.24.3
